requirements.txt
flask
pyjwt
cryptography
argon2-cffi
pytest
pytest-cov